Add path/alias import syntax, fix debug-tools disassembler hang, defensive opcode-mask cleanup
- import "path" [as name]: explicit relative/absolute path imports (../foo, /root/foo,
  C:/foo) alongside the existing dotted-identifier form, decoupling the bound name from
  the resolved file path (aer_module.c/h, parser.c).
- Fixed a real bug in AER_DEBUG_TOOLS's per-opcode summary pass (disasm.c): it masked
  opcodes with 0xFF instead of the codebase's established 0x7F, misdecoding any
  instruction whose adjacent packed field set bit 7 and looping forever with no further
  output — independent of runtime iteration count. Confirmed fixed at scale (sieve.aer
  LIMIT=200000 now completes in ~0.1s instead of hanging indefinitely).
- Applied the same 0x7F convention defensively to two OP_CALL/OP_TAIL_CALL opcode-rewrite
  sites in parser.c (safe today since those opcodes are PACK3-encoded with a dedicated
  opcode byte, but consistent with the rest of the codebase and guards against a future
  encoding change reintroducing the same bug class).
- Documented a newly-discovered, pre-existing Windows-only bug in README's Known Bugs
  section: vm_run's longjmp-based error unwind crashes via a failed SEH stack-unwind
  (STATUS_BAD_STACK) on any runtime error on Windows; confirmed unaffected on Linux/Pi via
  the full embedding test suite, LTO ruled out as the cause. Not fixed in this commit.
- New sieve.aer/.lua/.py three-way benchmark (matches the existing nbody.aer pattern).

All three test suites (test, test-embed, test-smoke) verified green on both Windows and
Raspberry Pi before this commit.

+++ source/compiler/lexer.c
@@ -149,7 +149,7 @@ static void emit(TokenType type, unsigned int length) {
     current->buffer += length;
 }
 
-static void emit_integer(long long integer, unsigned int length) {
+static void emit_integer(int64_t integer, unsigned int length) {
     token.value = aer_int(integer);
     emit(TOKEN_INTEGER, length);
 }
@@ -211,7 +211,7 @@ static void lex_number() {
             return;
         }
         char* end; errno = 0;
-        long long val = strtoll(buf, &end, 16);
+        int64_t val = strtoll(buf, &end, 16);
         if (errno == ERANGE) error_at("Integer literal overflow");
         if (isalpha((unsigned char)*end) || *end == '_')
             error_at("Invalid character after integer literal");
@@ -221,7 +221,7 @@ static void lex_number() {
 
     /* lex() only calls lex_number() on a digit, so int_len is always >= 1 — safe to fall through from error_at() below without an explicit return. */
     char* end; errno = 0;
-    long long int_val = strtoll(buf, &end, 10);
+    int64_t int_val = strtoll(buf, &end, 10);
     if (errno == ERANGE) error_at("Integer literal overflow");
     unsigned int int_len = (unsigned int)(end - buf);
